+++ src/kernel/coinstats.h
@@ -19,7 +19,7 @@ namespace node {
 class BlockManager;
 } // namespace node
 
-namespace node {
+namespace kernel {
 enum class CoinStatsHashType {
     HASH_SERIALIZED,
     MUHASH,
@@ -73,6 +73,6 @@ uint64_t GetBogoSize(const CScript& script_pub_key);
 CDataStream TxOutSer(const COutPoint& outpoint, const Coin& coin);
 
 std::optional<CCoinsStats> ComputeUTXOStats(CoinStatsHashType hash_type, CCoinsView* view, node::BlockManager& blockman, const std::function<void()>& interruption_point = {});
-} // namespace node
+} // namespace kernel
 
 #endif // BITCOIN_KERNEL_COINSTATS_H
