+++ src/validation.cpp
@@ -63,8 +63,8 @@ using node::BlockManager;
 using node::BlockMap;
 using node::CBlockIndexHeightOnlyComparator;
 using node::CBlockIndexWorkComparator;
-using node::CCoinsStats;
-using node::CoinStatsHashType;
+using kernel::CCoinsStats;
+using kernel::CoinStatsHashType;
 using node::fImporting;
 using node::fPruneMode;
 using node::fReindex;
